Moved query logic to another slice
