+++ src/features/query/querySelectors.ts
@@ -1,0 +1,4 @@
+import type { RootState } from '@/app/store';
+
+// Select tracklist query items
+export const selectTrackListQuery = (state: RootState) => state.query;
